smart_requirements.txt
numpy==1.26.4
pillow==10.1.0
moviepy==1.0.3
scikit-learn==1.3.2
insightface==0.7.3
onnxruntime>=1.14.0
